protect if filename is None.
svn path=/src/trunk/; revision=6062

+++ pootling/modules/pickleTM.py
@@ -48,5 +48,6 @@ class pickleTM:
             cPickle.dump(matcher, pickleFile)
             pickleFile.close()
         else:
-            os.remove(filename)
+            if not filename:
+                os.remove(filename)
